Need to revert ppsspp

+++ Update-RG351P.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 clear
 
-UPDATE_DATE="12192020"
+UPDATE_DATE="12192020-1"
 LOG_FILE="/home/ark/update$UPDATE_DATE.log"
 UPDATE_DONE="/home/ark/.config/.update$UPDATE_DATE"
 
@@ -800,7 +800,7 @@ if [ ! -f "/home/ark/.config/.update12182020" ]; then
 	touch "/home/ark/.config/.update12182020"
 fi
 
-if [ ! -f "$UPDATE_DONE" ]; then
+if [ ! -f "/home/ark/.config/.update12192020" ]; then
 
 	printf "\nUpdate lzdoom to swap OK and Cancel buttons\nUpdated PPSSPPSDL to version 1.10.3 version with batocera speedup\nCenter Solarus" | tee -a "$LOG_FILE"
 	sudo wget https://github.com/christianhaitian/arkos/raw/main/12192020/arkosupdate12192020.zip -O /home/ark/arkosupdate12192020.zip -a "$LOG_FILE"
@@ -821,7 +821,7 @@ if [ ! -f "$UPDATE_DONE" ]; then
 	sudo chown ark:ark -v /opt/solarus/libsolarus* | tee -a "$LOG_FILE"
 
 	if [ -f "/home/ark/.config/.update12192020" ]; then
-		printf "\nUpdate boot text to reflect current version of ArkOS\n" | tee -a "$LOG_FILE"
+	printf "\nUpdate boot text to reflect current version of ArkOS\n" | tee -a "$LOG_FILE"
 		sudo sed -i "/title\=/c\title\=ArkOS 1.5 ($UPDATE_DATE)" /usr/share/plymouth/themes/text.plymouth
 	else
 		printf "\nThe update couldn't complete because the package did not download correctly.\nPlease retry the update again." | tee -a "$LOG_FILE"
@@ -829,6 +829,31 @@ if [ ! -f "$UPDATE_DONE" ]; then
 		exit 1
 	fi
 
+	touch "/home/ark/.config/.update12192020"
+fi
+
+if [ ! -f "$UPDATE_DONE" ]; then
+
+	printf "\nRevert PPSSPP back to previous version...\n" | tee -a "$LOG_FILE"
+	sudo wget https://github.com/christianhaitian/arkos/raw/main/12192020/arkosupdate12192020-1.zip -O /home/ark/arkosupdate12192020-1.zip -a "$LOG_FILE"
+	if [ -f "/home/ark/arkosupdate12192020-1.zip" ]; then
+		sudo unzip -X -o /home/ark/arkosupdate12192020-1.zip -d / | tee -a "$LOG_FILE"
+		sudo rm -v /home/ark/arkosupdate12192020-1.zip | tee -a "$LOG_FILE"
+	else 
+		printf "\nThe update couldn't complete because the package did not download correctly.\nPlease retry the update again." | tee -a "$LOG_FILE"
+		echo $c_brightness > /sys/devices/platform/backlight/backlight/backlight/brightness
+		exit 1
+	fi
+
+	if [ -f "/home/ark/.config/.update12192020-1" ]; then
+	printf "\nUpdate boot text to reflect current version of ArkOS\n" | tee -a "$LOG_FILE"
+		sudo sed -i "/title\=/c\title\=ArkOS 1.5 ($UPDATE_DATE)" /usr/share/plymouth/themes/text.plymouth
+	else
+		printf "\nThe update couldn't complete because the package did not download correctly.\nPlease retry the update again." | tee -a "$LOG_FILE"
+		echo $c_brightness > /sys/devices/platform/backlight/backlight/backlight/brightness
+		exit 1
+	fi
+	
 	touch "$UPDATE_DONE"
 	rm -v -- "$0" | tee -a "$LOG_FILE"
 	printf "\033c" >> /dev/tty1
